IMPORTANT: watch API now has diffrent policy about when watcher callbacks are executed and which resources are given to them. The callbacks are executed when some of the watched resources are triggered AND some of the triggered resources are available. Only triggered AND available resources are given to the callback. As for callback execution at watch_lt, all the watched resources are considered to be triggered then. Added the test 19-fire.t to confirm this behavior.
The triggering behavior of select API remains the same (hopefully...). The callbacks are executed when some of the watched resources are triggered and some of the watched resources are available then. The set of provided resources is irrelevant of the set of triggered resources. So if you have resources A and B, both watched by a watcher and A is available and B is not available, the callback is executed and gets A even when you trigger only B. This is confusing.